Moved `Created Collection` API to it's own route and tweaked the UI
